Added Color By Option
Under the selection to choose what type of map background to use, we now have the option to color the circles by magnitude, year, depth, or duration.

There is some weird interaction between these two option boxes that needs to be fixed still. We also need to add a legend of some kind to let the user know what the colors mean.