Invisible hitbox for bars
Some of the bars are very small, so an invisible hitbox was added to each bar to make it easier to hover over it. Even when there is no bar due to no earthquakes matching the data, hovering over where the bar would be will still show the tooltip.

Also, when the user hovers over the bar, the bar pops up a little bit to better show that its selected.