Line chart tooltip
Made the tooltip the same format as the map and bar chart.

I also changed the color used on the line chart to something different than what is used for the magnitude data set in the map and bar chart. I chose teal, but we can change this further if needed.