Added brush to the bar chart
A brush was added to the bar chart. When the user brushes bars, the bar chart and the map update to show only the selected data.

The brushing is a little finicky with the invisible hitbox of the bars, so the user has to start the brush in between the bars and make sure the entire bar is included in the brush.

There is also a "reset brush" button to reset to the full dataset.

The brushing doesn't update the line chart yet because we still need to make the line chart using the same csv file as the other 2 visualizations.