Map Brush
Added a brush to the map. The brush allows users to focus on a specific area within the map.

The user can either be in pan mode or brush mode for the map. Both cannot be done simultaneously.

The brushes between the map and bar chart are separate, but both brushes effect the other visualization. So each brush can filter the data further.

There are also two "reset Brush" buttons that reset the brush for the map and bar chart individually.