Added line chart brush
Added a brush to the line chart.

Similar to how it worked when there wee just brushes on the map and bar chart, the user can use the line chart brush independently from the other 2 brushes. The burhs on the line chart will still update all 3 visualizations.

A Reset Time Brush button was added above the line chart to reset just the line chart brush. When clicked, all 3 visualizations will reset.

Also, the way the tick marks for the line chart was changed. Instead of just showing the years as tick marks. The ticks marks will dynamically change along with the brush to always show 5-6 ticks.